dev/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
dnspython==2.4.2
Faker==20.1.0
idna==3.5
pymongo==4.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
urllib3==2.1.0
